+++ Physics/vector4.cc
@@ -5,7 +5,7 @@ namespace PMG::Physics {
 
 	}
 
-	Vector4::Vector4(double x, double y, double z, double w) : x(x), y(y), z(z), w(w) {
+	Vector4::Vector4(float x, float y, float z, float w) : x(x), y(y), z(z), w(w) {
 
 	}
 }
